+++ DatadogCore/Private/ObjcAppLaunchHandler.m
@@ -4,11 +4,14 @@
 * Copyright 2019-Present Datadog, Inc.
 */
 
-#import <UIKit/UIKit.h>
 #import <pthread.h>
 #import <sys/sysctl.h>
 
 #import "ObjcAppLaunchHandler.h"
+
+#if TARGET_OS_IOS || TARGET_OS_TV
+#import <UIKit/UIKit.h>
+#endif
 
 // A very long application launch time is most-likely the result of a pre-warmed process.
 // We consider 30s as a threshold for pre-warm detection.
@@ -39,7 +42,7 @@ static __dd_private_AppLaunchHandler *_shared;
     // This is called at the `DatadogPrivate` load time, keep the work minimal
     _shared = [[self alloc] initWithProcessInfo:NSProcessInfo.processInfo
                                        loadTime:CFAbsoluteTimeGetCurrent()];
-
+#if TARGET_OS_IOS || TARGET_OS_TV
     NSNotificationCenter * __weak center = NSNotificationCenter.defaultCenter;
     id __block __unused token = [center addObserverForName:UIApplicationDidBecomeActiveNotification
                                                     object:nil
@@ -55,6 +58,7 @@ static __dd_private_AppLaunchHandler *_shared;
         [center removeObserver:token];
         token = nil;
     }];
+#endif
 }
 
 + (__dd_private_AppLaunchHandler *)shared {