+++ DatadogSessionReplay/Sources/SessionReplay.swift
@@ -47,6 +47,11 @@ public enum SessionReplay {
 
         let sessionReplay = try SessionReplayFeature(core: core, configuration: configuration)
         try core.register(feature: sessionReplay)
+
+        core.telemetry.configuration(
+            defaultPrivacyLevel: configuration.defaultPrivacyLevel.rawValue,
+            sessionReplaySampleRate: Int64(withNoOverflow: configuration.replaySampleRate)
+        )
     }
 }
 #endif